Do incremental pushes.
Change push to first retrieve the snapshots that exist remotely.
Supply this information to the local filesystem reader so it can construct an incremental stream.
Update the filesystem writer so it is capable of receiving incremental streams into the filesystem.
Push has become asynchronous - so propagate Deferred-handling code up the stack from it.