Add support for CoreOS production images
The previous instructions linked to a tarball whose toolchain would only
work on a CoreOS developer build and consequently, it was not possible
to build kernel modules using it when CoreOS updates put the modules and
kernel out of sync. This is remedied by modifying the toolchain to
include its own glibc. This enables us to build new modules on CoreOS
production images as needed.

Signed-off-by: Richard Yao <richard.yao@clusterhq.com>