Simplify handling of errors in remove.
New versions of docker return NOT_MODIFIED for stopped containers, when
stopping them, instead of INTERNAL_SERVER_ERROR. Since this isn't an error,
docker-py doesn't report the state.

(Note: We bump the version of the API that we are using to a version that
claims implements this. It turns out that a new enough docker *always* has
this behaviour).