Start of splitting up the Dockerfile-based package build into separate pieces, some of which are more cacheable or have cacheable artifacts.
Does not quite work though.  Fails to find `flocker` when it comes to actually do the install,
maybe because the wheel package is not installed and so pip cannot read wheel files?  Who knows.

top-level build.sh helper just until the pieces work and then packaging.py should be updated to
do this