Avoid potentially asynchronous work in calculate_necessary_state_changes.
Previously we looked up the volume list and used it to construct the list of delete changes necessary.
This will turn into an asynchronous operation real soon now.
Avoid this by moving the volume list operation into the change implementation.
This is horribly wasteful but we can fix it later by recording volume state as part of cluster state.
This will let calculate_necessary_state_changes only delete things that have recently been seen to exist.
It will also let it avoid doing a volume list at *all* as part of the deletion logic.

A cascade of follow-on changes are required to make this work.
Primarily, DestroyBlockDeviceDataset must operate on a dataset_id instead of a volume.
This removes the benefit of using proxyForInterface and changes an API used throughout the tests.