zfs: switch snapshot creation to lzc
Also, remove all unit tests for snapshots.
The unit tests were based on the fact that the operations were done via
zfs command, so they checked that the command was spawned with the correct
arguments and the output was correctly parsed.

To do:	invent the way to test the new implementation