zfs: switch filesystem sending and receiving to lzc
Note: lzc_send()/lzc_receive() do not close their end of the pipe
(the file descriptor, in general) when they are done writing to it.

Note: this API is synchrnous at the moment including a requirement
for a synchronous cleanup of the file descriptors.
There are functional tests to verify that.