Re-introduce pre-tag docs publication, but make it optional
It is useful to check that credentials are setup correctly, and
the basic layout of the S3 docs infrastructure is correct, before
doing the irreversable tag of the release.

As the release process does not use the output of these steps, it
is not technically required, and it may make sense to remove it from
a more automated release process.