zfs test_maximum_size_enforced: write the data in a stricter fashion
First, write the data in 8 KiB chunks to avoid allocating the whole
multi-megabyte object in memory.

Second, flush the data after writing exactly maximum_size.
This ensures that we do not get an error before the quota is exhausted.

Finally, write and flush one extra byte and check that it does trigger
the error.
assertRaises is used as a context manager now, so that we do not have to
define another function just to perform two calls under the assertion check.