add coverage tests for every job
update cobertura plugin so that it always generates report

The cobertura/coverage reports were not being generated when one of
the child builds would fail. As it was set to only include 'stable'
builds.

added a new key coverage_report

Not every job runs coverage, and produces coverage.xml reports.
To prevent jobs failing from lack of artifacts. there is now a new
coverage_report, which if set to true will add the cobertura
publisher to the job definition.