Make transient states ordered, to detect more races sooner.
If the volume is detached, and the transient states are
('in-use', 'detaching'), then once the 'detaching' state is reached,
the 'in-use' state is invalid for this transition, and would indicate that
the volume has been attached elsewhere.  Rather than lead to a timeout
error, make the 'in-use' state invalid once 'detaching' state is reached,
so this race is detected as an invalid state for this transition, which
provides more useful information than a timeout error.