fix test_resize_volume_invalid_max_size
First, resized_filesystem() callback was unused.
Second and more important, if the error-callback maximum_too_small()
was not called, e.g. because a backend did not try to resize
a volume at all, then the test would just pass.
Now we ensure that the resize operation fails with
MaximumSizeTooSmall exception.