Modify DestroyVolume to have a blockdevice_id field rather than a volume.
DestroyVolume only used the blockdevice_id field of the BlockDeviceVolume that
it was passed, and in fact was getting a stale value of attached_to in
DestroyBlockDeviceDataset.

Rather than incurring extra work to get the correct value of attached_to, change
DestroyVolume to only take a blockdevice_id instead of a full BlockDeviceVolume
with stale fields.