Add volumes to DeploymentState.
The volumes are of type BlockDeviceVolume. This exposes the current list of
volumes (and potentially more importantly the mapping from dataset_id to
blockdevice_id) to anything that has the cluster_state. The intention will be to
use this mapping in calculate_changes so that we can just pass the
blockdevice_id down to the IStateChanges so they don't have to do an API call to
get the blockdevice_id.