flocker-volume: ensure that the reactor is running
deferLater() unlike maybeDeferred ensures that the reactor is
actually running.  This is required by the new ZFS backend, because
it uses the reactor's thread pool to run lzc_send() and lzc_receive()
in separate threads.  That has to be done because those calls lend
their context to the kernel for doing the actual send / receive
work.  The main thread pumps the stream data between the userspace
and the kernel.