Change discover_state to return an ILocalState provider.
The ILocalState provider will be passed into calculate_changes, and facilitates
passing information from discover_state to calculate_changes without having to
add it to the DeployedState and transfer it to control agent. The ILocalState
interface has a method named shared_state_changes() that must return an iterable
of IClusterStateChange providers, which is the current return value from
discover_state.