Remove FullySharedLocalState in favor of NodeLocalState.
All real users of FullySharedLocalState only ever filled it with a single
NodeState. Rather than keeping an awkward class that stores a vector of
IClusterStateChanges, prefer to replace is with a simple class that has sane
structure.

NodeLocalState has proper structure rather than being a simple grab bag of
IClusterStateChanges.