Modify DestroyBlockDeviceDataset and DestroyVolume to take blockdevice_ids.
This enables the removal of the list_volumes call at the beginning of
DestroyBlockDeviceDataset.run, reducing total API calls.

While I was in the code I did a drive by fix to FLOC-1818, which was trivial.

Removed comment referencing  FLOC-1756 was already marked as 'Done' and
certainly the comment was no longer accurate.

Also removed comment referencing FLOC-1616, which was also already marked as
'Done', and the comment was no longer accurate.