add --purpose parameter to setup-cluster
The parameter's value gets recorded in the 'purpose' tag of the created
nodes if such metadata is possible to record.
The value is also used to build names of the created nodes.

For 'run-acceptance-tests' the purpose is hard-coded
as 'acceptance-testing'.
Please note that now the acceptance test nodes are named
"acceptance-testing-..." as opposed to "acceptance-test-..." used before.