setup-cluster: use completely random uuids for clusters
At this moment various cluster identification information is placed
into the root certificate that is generated for the cluster.
Specifically, into its CN and OU fields.
Thus, the certificate is a source of the identity information.

run-acceptance-tests uses flocker.provision.Certificates.generate() to
create certificates.

First, the method always used "acceptance-cluster" as a CN (commonName).
Now we allow the name to be set by a caller.

Second, the provided cluster ID (which is a UUID) was always specially
encoded and it had a characteristic "AAAA..." marker at the end.
Now we use that kind of a UUID only when we create an acceptance testing
cluster.

For setup-cluster we now use a CN derived from the purpose parameter and
a UUID that is completely random.