Rename jobs.groovy.j2 to jobs.groovy now it doesn't use jinja.
In preparation for dropping the jinja translation step
we rename jobs.groovy.j2 to jobs.groovy so that it would be
found by the setup_ClusterHQ-flocker job config.

A symlink with the old name is kept around so that there
doesn't have to be a co-ordinated change across all
branches with Jenkins.

Once jenkins has changed to not using jobs.groovy.j2 then
the symlink can be removed.