Move the copyArtifacts tasks back the multijob.
This reverts 31c881ccda449a3961896148f80093f853709c89 which moved
the copyArtifacts steps to the invidual jobs, rather than the multijob.

This means that after a build each job tries to copy the artifacts
from the previous run of the job to the current one. The intention
is to copy the results from the current build to the multijob so that
they can be examined in aggregate with the other jobs.

In order to do that the copy needs to be run in the multijob, as it's
"copy <files> to here" operation.

The original intent of the change was to have the artifacts published
for the cron jobs. That is in fact done by defining publishers, and
was done in a followup commit b245c5f7387285db60545f33d29336d2af78f3a2.