Refactor some of the groovy code to reduce duplication.
Rather than relying on constructing the same or similar strings
in several places use functions to encapsulate the details
to reduce the chance of a mismatch.

Unify some of the job definition as indicated by XXX comments.

Extract the code to trigger a sub job from a multijob without
failing the multijob in to a function.

Migrate from using "job" to trigger a sub job from a multijob
as that is deprecated in favour of using phaseJob. This
is currently showing the deprecation warning on the live
instance.