Increase timeout on node.functional.test_docker tests
Unfortunately this means increasing the timeout for
all trial tests, or defining new stanzas for just the
test_docker module.

test_docker is currently timing out fairly frequently
because the tests are slow to run. Increasing the
timeout will allow more green runs to happen (if slower)
while we look in to whether we can speed up the tests,
or parallelise them to get the complete suite faster.