Move the OSX installation test to run nightly.
The test is more of an acceptance test, and is also unlikely
to be broken by any particular change, so running it nightly
is reasonable.

In addition it is currently failing often, perhaps related
to parallel versions of the job, so running once a day will
avoid us having to diagnose the root cause of that.