Report to github the status of individual jobs rather than the multijob.
This means that the information on the github page will be more useful,
as you can see which tests failed without clicking through and
searching. In addition, when a job is retried it will update its status
again, so a passing run will show all green on github even if it
required a job to be retried.

This commit changes the "pending" notifications to be for each test
we intend to run, and then each job does the successful/failing
notifications itself.

The github_status.sh script with the auth details isn't available
on the slaves, so this now embeds it in the script step.