Create a function to expand the list of jobs.
A common pattern used in the file is to loop over the defined
jobs. However, they are split by category, and some categories
have different handling (sometimes there are modules, which
mean one job is in fact many, cron jobs use a slightly different
naming convention.)

This commit splits out a function to perform the loop,
expanding to the level of jobs, and calcualating the basic
info in once place.

Each call site can the filter the list and act on the jobs
that it wants without re-implementing that logic.

As well as some small cleanups this will also now raise an
error if build.yaml defines a job type that this script
doesn't know how to handle.