setup-cluster: re-organize error handling and cluster clean-up code
Because main() is decorated with @inlineCallbacks there is no guarantee
that the control flow always returns to it after a yield statement.
In particular, that's the case if the reactor is stopped.
And that could happen, for instance, if a user presses Ctrl-C because
twisted installs a SIGINT handler which stops the reactor.

Thus, to improve the chances of the cluster clean-up being executed
the clean-up code is moved to a before-shutdown system event trigger.

Also, there is no reason now to have except and finally clauses as
(a) there is no guarantee that they are executed, (b) all the except
block did was to re-raise an exception, so no reason to capture it
at all, (c) the finally block would contain only actions that should
be executed on success after the clean-up code was moved to the
shutdown handler.