Give test_docker on CentOS a larger slave and timeout.
docker on CentOS runs slow in the default configuration because
it uses loop-devmapper. It's quite hard for us to change that,
but the tests seem more reliable on a slightly larger instance
with a slightly longer timeout. Give it more resources so it
more consistently passes while we find another way to improve
the situation.