alternative implementation of LibcloudRunner.start_cluster in setup-cluster
The alternative implementation should be faster as it creates cloud nodes
in parallel.  Also it should be more robust as it does not fail as long
as it is able to create at least a control node for the cluster.

This change results in a bit of code duplication between the old and the
new implementations.  That is supposed to be only a temporary situation.
We do not want to make big changes to the acceptance testing code
at this moment.