combine create and provision operations and retry the compound operation
In the most well known case of the failed provisioning it does not make
sense to retry just the provisioning as it would always fail unless
the affected node is rebooted or re-created.  That's the case where
the cloud-init mysteriously fails to fix up the sudoers file and we can not
sudo via ssh.
At this moment I opt for re-creating the node, but it's worth while
exploring the option of rebooting the node.