use addCallbacks over addErrback + addCallback if errback can not recover
This should be a nop-change, because the involved errbacks pass through
the failure, but the new code should make it easier to see that
the callbacks won't run if there is a failure.

Also, use more descriptive names for the errbacks, so that it is easier
to see what failures they handle.