Add a script that can help with merging a pull request.
admin/merge_pr https://github.com/ClusterHQ/flocker/pull/<id>

will loop until all the checks are green, then trigger a merge
and delete the branch.

Rather than having to decide between merging with failing tests,
or spending time triggering retries and checking status, you can
use this script to do the work. Give it the URL of the pull request
and it will loop checking the status. When it is green it will
merge the PR, delete the source branch from github and exit.

If tests fail then it will retry the Jenkins job once, and if they
fail again it will tell you about it so that you can take a look.
If the test failures aren't your fault you can click retry and it
will merge if it is green from that run. If they are your fault
then kill the script and make the fixes.

If you just want to get a green run before asking for a review
then use --no-merge, it will exit when the tests are green without
triggering the merge.

You could combine the script with "undistract-me" or "notify-send"
to get a popup when the tests are green.