IProvisioner.create_nodes: a method for efficiently creating multiple nodes
Unlike create_node() this method is asynchronous and returns a list of
deferreds.
With this commit the method is implemented in the AWS provisioner.
AWS (and boto) API allows to request multiple instances at once and then
we check states of each individual instance in parallel, using the
reactor's loop.