Harden flocker-dataset-id EBS parsing
If there is a volume in EBS with a tag key of `flocker-dataset-id` but
the value is not a valid UUID, any Flocker agent connected to that
account will throw an exception during iteration and will be unusable.

The change here catches bad values and makes them process as just
another volume that's not part of the asking cluster rather than
throwing the mentioned exceptions.